requirements.txt
arm-mango==1.2.0
conda==22.9.0
h5py==3.7.0
keras==2.10.0
Keras-Preprocessing==1.1.2
matplotlib==3.6.0
matplotlib-inline==0.1.6
numpy==1.23.3
opencv-python==4.6.0.66
pandas==1.5.0
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
tqdm